Add new plpgsql_check.profiler_max_shared_chunks GUC.
As seen in https://github.com/okbob/plpgsql_check/issues/81 this commit add a
new GUC to allow users to change the previously hardcoded MAX_SHARED_CHUNKS.

The GUC is defined as PGC_POSTMASTER, as it has to be set at start time, but is
only defined when plpgsql_check is loaded via shared_preload_libraries, as it
would otherwise prevent users from loading the extension interactively.

The default values of 15000 is kept, and value between 50 and 100000 are
allowed.