An plpgsql compiler try to cache compiled code, and should to returned fresh compiled code, when function was not compiled in session or when function was changed, or returns cached compiled code (pointer to statemnt tree). I expected so this statement tree can be only one for unchanged function. But it is not true. There can be more different (not identical) trees in dependency on types of arguments and some details of execution.
That is an problem, because I compared pointers of statements to get
statement id stored in statement map. But statement map generated for
one statement tree cannot be used for searching statementid of statement
of different statement tree.

Now, I hold an list of statement trees with generated statement map. When
tree is not in this list, then current statement map is enhanced about
statements from this tree. Entry of statement map has now two fields:
pointer to function, and pointer to statement.

Second fixed issue was generating some items of statement map in short
life memory context. Any field of statement map should to be created
in profiler_mcxt. The prerequisity for this currupted statement map
was multiple statements on one line.