Properly ignore syntax errors from placeholders
The old code tried to skip parsing when there are any string literals in
the dynamic SQL but due to a suspected typo (false instead of true) it
did not actually do so.

An example where literal placeholders can cause syntax errors in valid
code is format('SELECT json %L', v) which check_dynamic_sql() transforms
into "SELECT json null" which is invalid SQL in PostgreSQL <=13.