Fix false positive for record return type
PostgreSQL for some reason instantiates the anonymous record return
type for functions and this causes the record structure comparison to
fail in some instances.

And since it does not really make sense to compare the structure of
a returned tuple with a fully anonymous return type (they should
always match) we may as well skip the check for this case instead of
trying to fix it.