release plans created by plpgsql_check
in passive mode. Without it, the plan can be inconsistent for plpgsql and some
plpgsql assertions.

the trackers tests was executed with enabled passive check. Previously, the plans
created by passive check was not removed, and it did impacts on deep of contexts
and result of tracking. This patch fixes output of regress tests, and disable
passive check mode before tracker tests.

related to #174